Move parsePseudoElement() to CSSSelectorParser https://bugs.webkit.org/show_bug.cgi?id=269597
Reviewed by Tim Nguyen.

Per review feedback in 274846@main. Also elaborate the FIXME to better
explain why consolidation with consumePseudo() is non-trivial and
perhaps not worth it until we have substantially more pseudo elements
with arguments.

* Source/WebCore/animation/WebAnimationUtilities.cpp:
(WebCore::pseudoIdFromString):
* Source/WebCore/css/CSSSelector.cpp:
(WebCore::pseudoElementIdentifierFor): Deleted.
(WebCore::CSSSelector::parsePseudoElement): Deleted.
* Source/WebCore/css/CSSSelector.h:
* Source/WebCore/css/parser/CSSSelectorParser.cpp:
(WebCore::pseudoElementIdentifierFor):
(WebCore::CSSSelectorParser::parsePseudoElement):
* Source/WebCore/css/parser/CSSSelectorParser.h:
* Source/WebCore/page/LocalDOMWindow.cpp:
(WebCore::LocalDOMWindow::getComputedStyle const):
(WebCore::LocalDOMWindow::getMatchedCSSRules const):

Canonical link: https://commits.webkit.org/274893@main
